fix `admin` crate dependency on `storage-query` flag
Summary:
cluster controller had a dependency on a non default `storage-query` feature.

Fixes #3002
