+++ crates/admin/src/lib.rs
@@ -12,6 +12,7 @@ pub mod cluster_controller;
 mod error;
 #[cfg(feature = "metadata-api")]
 mod metadata_api;
+mod query_utils;
 mod rest_api;
 mod schema_registry;
 pub mod service;
