+++ crates/admin/src/storage_query/mod.rs
@@ -15,7 +15,6 @@ mod query;
 use axum::{Router, routing::post};
 use std::sync::Arc;
 
-pub use query::{RecordBatchWriter, WriteRecordBatchStream};
 use restate_storage_query_datafusion::context::QueryContext;
 
 #[derive(Clone)]
